+++ src/settings/OrderTemplates/OrderTemplatesEditor/OrderTemplatesEditor.test.js
@@ -20,6 +20,10 @@ jest.mock('@folio/stripes/components', () => ({
   expandAllSections: jest.fn(),
   Layer: jest.fn(({ children }) => <>{children}</>),
 }));
+jest.mock('@folio/stripes/smart-components', () => ({
+  ...jest.requireActual('@folio/stripes/smart-components'),
+  EditCustomFieldsRecord: jest.fn().mockReturnValue('EditCustomFieldsRecord'),
+}));
 
 const defaultProps = {
   close: jest.fn(),
@@ -110,6 +114,7 @@ describe('OrderTemplatesEditor', () => {
       expect(screen.getByText('ui-orders.settings.orderTemplates.accordion.polFundDistribution')).toBeInTheDocument();
       expect(screen.getByText('ui-orders.settings.orderTemplates.accordion.polLocation')).toBeInTheDocument();
       expect(screen.getByText('ui-orders.settings.orderTemplates.accordion.polTags')).toBeInTheDocument();
+      expect(screen.getByText('EditCustomFieldsRecord')).toBeInTheDocument();
     });
   });